Migrate device-auth modules to live inside autofill-impl
