+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/deviceauth/DeviceAuthenticator.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 DuckDuckGo
+ * Copyright (c) 2023 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,59 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.deviceauth.impl
+package com.duckduckgo.autofill.impl.deviceauth
 
 import android.os.Build
 import androidx.annotation.UiThread
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import com.duckduckgo.appbuildconfig.api.AppBuildConfig
-import com.duckduckgo.deviceauth.api.AutofillAuthorizationGracePeriod
-import com.duckduckgo.deviceauth.api.DeviceAuthenticator
-import com.duckduckgo.deviceauth.api.DeviceAuthenticator.AuthResult
-import com.duckduckgo.deviceauth.api.DeviceAuthenticator.Features
+import com.duckduckgo.autofill.impl.R
+import com.duckduckgo.autofill.impl.deviceauth.DeviceAuthenticator.AuthResult
+import com.duckduckgo.autofill.impl.deviceauth.DeviceAuthenticator.Features
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
+
+interface DeviceAuthenticator {
+    /**
+     * This method can be used to check if the user's device has a valid device authentication enrolled (Fingerprint, PIN, pattern or password).
+     */
+    fun hasValidDeviceAuthentication(): Boolean
+
+    /**
+     * Launches a device authentication flow for a specific [featureToAuth] from a [fragment]. [onResult] can be used to
+     * communicate back to the feature the result of the flow.
+     */
+    @UiThread
+    fun authenticate(
+        featureToAuth: Features,
+        fragment: Fragment,
+        onResult: (AuthResult) -> Unit,
+    )
+
+    /**
+     * Launches a device authentication flow for a specific [featureToAuth] from a [fragmentActivity]. [onResult] can be used to
+     * communicate back to the feature the result of the flow.
+     */
+    @UiThread
+    fun authenticate(
+        featureToAuth: Features,
+        fragmentActivity: FragmentActivity,
+        onResult: (AuthResult) -> Unit,
+    )
+
+    sealed class AuthResult {
+        object Success : AuthResult()
+        object UserCancelled : AuthResult()
+        data class Error(val reason: String) : AuthResult()
+    }
+
+    enum class Features {
+        AUTOFILL_TO_USE_CREDENTIALS,
+        AUTOFILL_TO_ACCESS_CREDENTIALS,
+    }
+}
 
 @ContributesBinding(AppScope::class)
 class RealDeviceAuthenticator @Inject constructor(
