+++ autofill/autofill-impl/src/main/res/values-el/strings-device-auth.xml
