+++ autofill/autofill-impl/src/main/res/values-et/strings-device-auth.xml
