+++ autofill/autofill-impl/src/main/res/values-fr/strings-device-auth.xml
