+++ autofill/autofill-impl/src/main/res/values-lt/strings-device-auth.xml
