+++ autofill/autofill-impl/src/main/res/values-lv/strings-device-auth.xml
