+++ autofill/autofill-impl/src/main/res/values-nb/strings-device-auth.xml
