--- device-auth/device-auth-api/.gitignore
@@ -1,1 +1,0 @@
-/build
