--- device-auth/device-auth-api/src/main/java/com/duckduckgo/deviceauth/api/AutofillAuthorizationGracePeriod.kt
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2022 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.deviceauth.api
-
-/**
- * A grace period for autofill authorization.
- * This is used to allow autofill authorization to be skipped for a short period of time after a successful authorization.
- */
-interface AutofillAuthorizationGracePeriod {
-
-    /**
-     * Can be used to determine if device auth is required. If not required, it can be bypassed.
-     * @return true if authorization is required, false otherwise
-     */
-    fun isAuthRequired(): Boolean
-
-    /**
-     * Records the timestamp of a successful device authorization
-     */
-    fun recordSuccessfulAuthorization()
-
-    /**
-     * Invalidates the grace period, so that the next call to [isAuthRequired] will return true
-     */
-    fun invalidate()
-}
