--- device-auth/device-auth-impl/.gitignore
@@ -1,1 +1,0 @@
-/build
