--- device-auth/device-auth-impl/build.gradle
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2022 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'com.squareup.anvil'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-dependencies {
-    implementation project(path: ':app-build-config-api')
-    implementation project(path: ':device-auth-api')
-    implementation project(path: ':di')
-    implementation AndroidX.biometric
-    implementation Google.dagger
-    implementation JakeWharton.timber
-
-    testImplementation Testing.junit4
-    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
-    testImplementation project(path: ':common-test')
-}
-
-android {
-    anvil {
-        generateDaggerFactories = true // default is false
-    }
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
-    }
-  namespace 'com.duckduckgo.deviceauth.impl'
-}
